Convert exercise length to long for timer support

+++ app/src/main/java/com/example/musictoned/workoutcreation/WorkoutExercise.kt
@@ -3,7 +3,7 @@ package com.example.musictoned.workoutcreation
 //input is workout name, time in seconds
 data class WorkoutExercise(
     private val workout: Exercise,
-    private var length: Int = 30,
+    private var length: Long = 30,
     private var songSetting: SongSetting = SongSetting.SHUFFLE,
     private var bpmMode: BpmMode = BpmMode.AVERAGE,
     private var song: String = "" //song name in string? depends on what info needed to play the song
@@ -12,7 +12,7 @@ data class WorkoutExercise(
         return workout
     }
 
-    fun getLength(): Int{
+    fun getLength(): Long{
         return length
     }
